+++ src/app/campaign/store/models/combined.models.ts
@@ -1,0 +1,9 @@
+import { Campaign } from './campaign.models';
+import { Flight } from './flight.models';
+
+export interface CampaignFormSave {
+  campaign: Campaign;
+  updatedFlights: Flight[];
+  createdFlights: Flight[];
+  deletedFlights: Flight[];
+}
