+++ src/app/campaign/store/models/haldoc.utils.ts
@@ -1,0 +1,7 @@
+import { HalDoc } from 'ngx-prx-styleguide';
+
+export const filterUnderscores = (doc: HalDoc): {} => {
+  return Object.keys(doc.asJSON())
+    .filter(key => !key.startsWith('_'))
+    .reduce((obj, key) => ({ ...obj, [key]: doc[key] }), {});
+};
