+++ src/app/campaign/store/models/index.ts
@@ -1,1 +1,3 @@
 export * from './campaign.models';
+export * from './flight.models';
+export * from './combined.models';
